+++ structure.sql
@@ -1,10 +1,10 @@
 CREATE TABLE IF NOT EXISTS `user` (
-  `id` bigint COMMENT 'Unique user identifier',
-  `is_bot` tinyint(1) DEFAULT 0 COMMENT 'True if this user is a bot',
-  `first_name` CHAR(255) NOT NULL DEFAULT '' COMMENT 'User''s first name',
-  `last_name` CHAR(255) DEFAULT NULL COMMENT 'User''s last name',
-  `username` CHAR(191) DEFAULT NULL COMMENT 'User''s username',
-  `language_code` CHAR(10) DEFAULT NULL COMMENT 'User''s system language',
+  `id` bigint COMMENT 'Unique identifier for this user or bot',
+  `is_bot` tinyint(1) DEFAULT 0 COMMENT 'True, if this user is a bot',
+  `first_name` CHAR(255) NOT NULL DEFAULT '' COMMENT 'User''s or bot''s first name',
+  `last_name` CHAR(255) DEFAULT NULL COMMENT 'User''s or bot''s last name',
+  `username` CHAR(191) DEFAULT NULL COMMENT 'User''s or bot''s username',
+  `language_code` CHAR(10) DEFAULT NULL COMMENT 'IETF language tag of the user''s language',
   `created_at` timestamp NULL DEFAULT NULL COMMENT 'Entry date creation',
   `updated_at` timestamp NULL DEFAULT NULL COMMENT 'Entry date update',
 
@@ -13,10 +13,12 @@ CREATE TABLE IF NOT EXISTS `user` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 CREATE TABLE IF NOT EXISTS `chat` (
-  `id` bigint COMMENT 'Unique user or chat identifier',
-  `type` ENUM('private', 'group', 'supergroup', 'channel') NOT NULL COMMENT 'Chat type, either private, group, supergroup or channel',
-  `title` CHAR(255) DEFAULT '' COMMENT 'Chat (group) title, is null if chat type is private',
+  `id` bigint COMMENT 'Unique identifier for this chat',
+  `type` ENUM('private', 'group', 'supergroup', 'channel') NOT NULL COMMENT 'Type of chat, can be either private, group, supergroup or channel',
+  `title` CHAR(255) DEFAULT '' COMMENT 'Title, for supergroups, channels and group chats',
   `username` CHAR(255) DEFAULT NULL COMMENT 'Username, for private chats, supergroups and channels if available',
+  `first_name` CHAR(255) DEFAULT NULL COMMENT 'First name of the other party in a private chat',
+  `last_name` CHAR(255) DEFAULT NULL COMMENT 'Last name of the other party in a private chat',
   `all_members_are_administrators` tinyint(1) DEFAULT 0 COMMENT 'True if a all members of this group are admins',
   `created_at` timestamp NULL DEFAULT NULL COMMENT 'Entry date creation',
   `updated_at` timestamp NULL DEFAULT NULL COMMENT 'Entry date update',
@@ -52,9 +54,9 @@ CREATE TABLE IF NOT EXISTS `inline_query` (
 
 CREATE TABLE IF NOT EXISTS `chosen_inline_result` (
   `id` bigint UNSIGNED AUTO_INCREMENT COMMENT 'Unique identifier for this entry',
-  `result_id` CHAR(255) NOT NULL DEFAULT '' COMMENT 'Identifier for this result',
-  `user_id` bigint NULL COMMENT 'Unique user identifier',
-  `location` CHAR(255) NULL DEFAULT NULL COMMENT 'Location object, user''s location',
+  `result_id` CHAR(255) NOT NULL DEFAULT '' COMMENT 'The unique identifier for the result that was chosen',
+  `user_id` bigint NULL COMMENT 'The user that chose the result',
+  `location` CHAR(255) NULL DEFAULT NULL COMMENT 'Sender location, only for bots that require user location',
   `inline_message_id` CHAR(255) NULL DEFAULT NULL COMMENT 'Identifier of the sent inline message',
   `query` TEXT NOT NULL COMMENT 'The query that was used to obtain the result',
   `created_at` timestamp NULL DEFAULT NULL COMMENT 'Entry date creation',
@@ -63,16 +65,6 @@ CREATE TABLE IF NOT EXISTS `chosen_inline_result` (
   KEY `user_id` (`user_id`),
 
   FOREIGN KEY (`user_id`) REFERENCES `user` (`id`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
-
-CREATE TABLE IF NOT EXISTS `poll` (
-  `id` bigint UNSIGNED COMMENT 'Unique poll identifier',
-  `question` char(255) NOT NULL COMMENT 'Poll question',
-  `options` text NOT NULL COMMENT 'List of poll options',
-  `is_closed` tinyint(1) DEFAULT 0 COMMENT 'True, if the poll is closed',
-  `created_at` timestamp NULL DEFAULT NULL COMMENT 'Entry date creation',
-
-  PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 CREATE TABLE IF NOT EXISTS `message` (
@@ -88,9 +80,12 @@ CREATE TABLE IF NOT EXISTS `message` (
   `forward_date` timestamp NULL DEFAULT NULL COMMENT 'date the original message was sent in timestamp format',
   `reply_to_chat` bigint NULL DEFAULT NULL COMMENT 'Unique chat identifier',
   `reply_to_message` bigint UNSIGNED DEFAULT NULL COMMENT 'Message that this message is reply to',
+  `edit_date` bigint UNSIGNED DEFAULT NULL COMMENT 'Date the message was last edited in Unix time',
   `media_group_id` TEXT COMMENT 'The unique identifier of a media message group this message belongs to',
+  `author_signature` TEXT COMMENT 'Signature of the post author for messages in channels',
   `text` TEXT COMMENT 'For text messages, the actual UTF-8 text of the message max message length 4096 char utf8mb4',
   `entities` TEXT COMMENT 'For text messages, special entities like usernames, URLs, bot commands, etc. that appear in the text',
+  `caption_entities` TEXT COMMENT 'For messages with a caption, special entities like usernames, URLs, bot commands, etc. that appear in the caption',
   `audio` TEXT COMMENT 'Audio object. Message is an audio file, information about the file',
   `document` TEXT COMMENT 'Document object. Message is a general file, information about the file',
   `animation` TEXT COMMENT 'Message is an animation, information about the animation',
@@ -100,11 +95,11 @@ CREATE TABLE IF NOT EXISTS `message` (
   `video` TEXT COMMENT 'Video object. Message is a video, information about the video',
   `voice` TEXT COMMENT 'Voice Object. Message is a Voice, information about the Voice',
   `video_note` TEXT COMMENT 'VoiceNote Object. Message is a Video Note, information about the Video Note',
+  `caption` TEXT COMMENT  'For message with caption, the actual UTF-8 text of the caption',
   `contact` TEXT COMMENT 'Contact object. Message is a shared contact, information about the contact',
   `location` TEXT COMMENT 'Location object. Message is a shared location, information about the location',
   `venue` TEXT COMMENT 'Venue object. Message is a Venue, information about the Venue',
   `poll` TEXT COMMENT 'Poll object. Message is a native poll, information about the poll',
-  `caption` TEXT COMMENT  'For message with caption, the actual UTF-8 text of the caption',
   `new_chat_members` TEXT COMMENT 'List of unique user identifiers, new member(s) were added to the group, information about them (one of these members may be the bot itself)',
   `left_chat_member` bigint NULL DEFAULT NULL COMMENT 'Unique user identifier, a member was removed from the group, information about them (this member may be the bot itself)',
   `new_chat_title` CHAR(255) DEFAULT NULL COMMENT 'A chat title was changed to this value',
@@ -116,6 +111,8 @@ CREATE TABLE IF NOT EXISTS `message` (
   `migrate_to_chat_id` bigint NULL DEFAULT NULL COMMENT 'Migrate to chat identifier. The group has been migrated to a supergroup with the specified identifier',
   `migrate_from_chat_id` bigint NULL DEFAULT NULL COMMENT 'Migrate from chat identifier. The supergroup has been migrated from a group with the specified identifier',
   `pinned_message` TEXT NULL COMMENT 'Message object. Specified message was pinned',
+  `invoice` TEXT NULL COMMENT 'Message is an invoice for a payment, information about the invoice',
+  `successful_payment` TEXT NULL COMMENT 'Message is a service message about a successful payment, information about the payment',
   `connected_website` TEXT NULL COMMENT 'The domain name of the website on which the user has logged in.',
   `passport_data` TEXT NULL COMMENT 'Telegram Passport data',
 
@@ -138,24 +135,6 @@ CREATE TABLE IF NOT EXISTS `message` (
   FOREIGN KEY (`left_chat_member`) REFERENCES `user` (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
-CREATE TABLE IF NOT EXISTS `callback_query` (
-  `id` bigint UNSIGNED COMMENT 'Unique identifier for this query',
-  `user_id` bigint NULL COMMENT 'Unique user identifier',
-  `chat_id` bigint NULL COMMENT 'Unique chat identifier',
-  `message_id` bigint UNSIGNED COMMENT 'Unique message identifier',
-  `inline_message_id` CHAR(255) NULL DEFAULT NULL COMMENT 'Identifier of the message sent via the bot in inline mode, that originated the query',
-  `data` CHAR(255) NOT NULL DEFAULT '' COMMENT 'Data associated with the callback button',
-  `created_at` timestamp NULL DEFAULT NULL COMMENT 'Entry date creation',
-
-  PRIMARY KEY (`id`),
-  KEY `user_id` (`user_id`),
-  KEY `chat_id` (`chat_id`),
-  KEY `message_id` (`message_id`),
-
-  FOREIGN KEY (`user_id`) REFERENCES `user` (`id`),
-  FOREIGN KEY (`chat_id`, `message_id`) REFERENCES `message` (`chat_id`, `id`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
-
 CREATE TABLE IF NOT EXISTS `edited_message` (
   `id` bigint UNSIGNED AUTO_INCREMENT COMMENT 'Unique identifier for this entry',
   `chat_id` bigint COMMENT 'Unique chat identifier',
@@ -176,27 +155,101 @@ CREATE TABLE IF NOT EXISTS `edited_message` (
   FOREIGN KEY (`user_id`) REFERENCES `user` (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
+CREATE TABLE IF NOT EXISTS `callback_query` (
+  `id` bigint UNSIGNED COMMENT 'Unique identifier for this query',
+  `user_id` bigint NULL COMMENT 'Unique user identifier',
+  `chat_id` bigint NULL COMMENT 'Unique chat identifier',
+  `message_id` bigint UNSIGNED COMMENT 'Unique message identifier',
+  `inline_message_id` CHAR(255) NULL DEFAULT NULL COMMENT 'Identifier of the message sent via the bot in inline mode, that originated the query',
+  `chat_instance` CHAR(255) NOT NULL DEFAULT '' COMMENT 'Global identifier, uniquely corresponding to the chat to which the message with the callback button was sent',
+  `data` CHAR(255) NOT NULL DEFAULT '' COMMENT 'Data associated with the callback button',
+  `game_short_name` CHAR(255) NOT NULL DEFAULT '' COMMENT 'Short name of a Game to be returned, serves as the unique identifier for the game',
+  `created_at` timestamp NULL DEFAULT NULL COMMENT 'Entry date creation',
+
+  PRIMARY KEY (`id`),
+  KEY `user_id` (`user_id`),
+  KEY `chat_id` (`chat_id`),
+  KEY `message_id` (`message_id`),
+
+  FOREIGN KEY (`user_id`) REFERENCES `user` (`id`),
+  FOREIGN KEY (`chat_id`, `message_id`) REFERENCES `message` (`chat_id`, `id`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
+
+CREATE TABLE IF NOT EXISTS `shipping_query` (
+  `id` bigint UNSIGNED COMMENT 'Unique query identifier',
+  `user_id` bigint COMMENT 'User who sent the query',
+  `invoice_payload` CHAR(255) NOT NULL DEFAULT '' COMMENT 'Bot specified invoice payload',
+  `shipping_address` CHAR(255) NOT NULL DEFAULT '' COMMENT 'User specified shipping address',
+  `created_at` timestamp NULL DEFAULT NULL COMMENT 'Entry date creation',
+
+  PRIMARY KEY (`id`),
+  KEY `user_id` (`user_id`),
+
+  FOREIGN KEY (`user_id`) REFERENCES `user` (`id`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
+
+CREATE TABLE IF NOT EXISTS `pre_checkout_query` (
+  `id` bigint UNSIGNED COMMENT 'Unique query identifier',
+  `user_id` bigint COMMENT 'User who sent the query',
+  `currency` CHAR(3) COMMENT 'Three-letter ISO 4217 currency code',
+  `total_amount` bigint COMMENT 'Total price in the smallest units of the currency',
+  `invoice_payload` CHAR(255) NOT NULL DEFAULT '' COMMENT 'Bot specified invoice payload',
+  `shipping_option_id` CHAR(255) NULL COMMENT 'Identifier of the shipping option chosen by the user',
+  `order_info` TEXT NULL COMMENT 'Order info provided by the user',
+  `created_at` timestamp NULL DEFAULT NULL COMMENT 'Entry date creation',
+
+  PRIMARY KEY (`id`),
+  KEY `user_id` (`user_id`),
+
+  FOREIGN KEY (`user_id`) REFERENCES `user` (`id`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
+
+CREATE TABLE IF NOT EXISTS `poll` (
+  `id` bigint UNSIGNED COMMENT 'Unique poll identifier',
+  `question` char(255) NOT NULL COMMENT 'Poll question',
+  `options` text NOT NULL COMMENT 'List of poll options',
+  `is_closed` tinyint(1) DEFAULT 0 COMMENT 'True, if the poll is closed',
+  `created_at` timestamp NULL DEFAULT NULL COMMENT 'Entry date creation',
+
+  PRIMARY KEY (`id`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
+
 CREATE TABLE IF NOT EXISTS `telegram_update` (
   `id` bigint UNSIGNED COMMENT 'Update''s unique identifier',
   `chat_id` bigint NULL DEFAULT NULL COMMENT 'Unique chat identifier',
-  `message_id` bigint UNSIGNED DEFAULT NULL COMMENT 'Unique message identifier',
-  `inline_query_id` bigint UNSIGNED DEFAULT NULL COMMENT 'Unique inline query identifier',
-  `chosen_inline_result_id` bigint UNSIGNED DEFAULT NULL COMMENT 'Local chosen inline result identifier',
-  `callback_query_id` bigint UNSIGNED DEFAULT NULL COMMENT 'Unique callback query identifier',
-  `edited_message_id` bigint UNSIGNED DEFAULT NULL COMMENT 'Local edited message identifier',
+  `message_id` bigint UNSIGNED DEFAULT NULL COMMENT 'New incoming message of any kind - text, photo, sticker, etc.',
+  `edited_message_id` bigint UNSIGNED DEFAULT NULL COMMENT 'New version of a message that is known to the bot and was edited',
+  `channel_post_id` bigint UNSIGNED DEFAULT NULL COMMENT 'New incoming channel post of any kind - text, photo, sticker, etc.',
+  `edited_channel_post_id` bigint UNSIGNED DEFAULT NULL COMMENT 'New version of a channel post that is known to the bot and was edited',
+  `inline_query_id` bigint UNSIGNED DEFAULT NULL COMMENT 'New incoming inline query',
+  `chosen_inline_result_id` bigint UNSIGNED DEFAULT NULL COMMENT 'The result of an inline query that was chosen by a user and sent to their chat partner',
+  `callback_query_id` bigint UNSIGNED DEFAULT NULL COMMENT 'New incoming callback query',
+  `shipping_query_id` bigint UNSIGNED DEFAULT NULL COMMENT 'New incoming shipping query. Only for invoices with flexible price',
+  `pre_checkout_query_id` bigint UNSIGNED DEFAULT NULL COMMENT 'New incoming pre-checkout query. Contains full information about checkout',
+  `poll_id` bigint UNSIGNED DEFAULT NULL COMMENT 'New poll state. Bots receive only updates about polls, which are sent or stopped by the bot',
 
   PRIMARY KEY (`id`),
   KEY `message_id` (`chat_id`, `message_id`),
+  KEY `edited_message_id` (`edited_message_id`),
+  KEY `channel_post_id` (`channel_post_id`),
+  KEY `edited_channel_post_id` (`edited_channel_post_id`),
   KEY `inline_query_id` (`inline_query_id`),
   KEY `chosen_inline_result_id` (`chosen_inline_result_id`),
   KEY `callback_query_id` (`callback_query_id`),
-  KEY `edited_message_id` (`edited_message_id`),
+  KEY `shipping_query_id` (`shipping_query_id`),
+  KEY `pre_checkout_query_id` (`pre_checkout_query_id`),
+  KEY `poll_id` (`poll_id`),
 
   FOREIGN KEY (`chat_id`, `message_id`) REFERENCES `message` (`chat_id`, `id`),
+  FOREIGN KEY (`edited_message_id`) REFERENCES `edited_message` (`id`),
+  FOREIGN KEY (`chat_id`, `channel_post_id`) REFERENCES `message` (`chat_id`, `id`),
+  FOREIGN KEY (`edited_channel_post_id`) REFERENCES `edited_message` (`id`),
   FOREIGN KEY (`inline_query_id`) REFERENCES `inline_query` (`id`),
   FOREIGN KEY (`chosen_inline_result_id`) REFERENCES `chosen_inline_result` (`id`),
   FOREIGN KEY (`callback_query_id`) REFERENCES `callback_query` (`id`),
-  FOREIGN KEY (`edited_message_id`) REFERENCES `edited_message` (`id`)
+  FOREIGN KEY (`shipping_query_id`) REFERENCES `shipping_query` (`id`),
+  FOREIGN KEY (`pre_checkout_query_id`) REFERENCES `pre_checkout_query` (`id`),
+  FOREIGN KEY (`poll_id`) REFERENCES `poll` (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_520_ci;
 
 CREATE TABLE IF NOT EXISTS `conversation` (